Moved the consistent column defs and the estimated hours column def into columns.jsx
They no longer need any extra props passed down to them
